Reserve x20 for JitContext pointer, eliminate TLS in inline IC
- Reserve x20 (callee-saved) for JitContext pointer on aarch64.
  Set at all interpreter→JIT entry points via ensure_ctx_reg().
  Survives across all function calls (callee-saved ABI).

- x19 is reserved by LLVM on Apple platforms; use x20 instead.

- Replace wren_ic_enter/leave CallRuntime (function call + TLS access)
  with direct loads/stores through x20 in the kind=6 inline IC path:
  ldr saved, [x20, #40]  ; save current_func_id
  str new_id, [x20, #40] ; set new func_id
  str closure, [x20, #48]; set closure
  blr jit_fn              ; direct call
  str saved, [x20, #40]  ; restore

- Widen JitContext.current_func_id from u32 to u64 for clean
  8-byte aligned access from JIT code.

Infrastructure for future optimizations that need VM/context access
from JIT code without TLS (e.g., inline field setters, allocation).

+++ src/runtime/engine.rs
@@ -1439,7 +1439,7 @@ mod tests {
                 vm: dummy_vm.as_mut_ptr(),
                 module_name: std::ptr::null(),
                 module_name_len: 0,
-                current_func_id: u32::MAX,
+                current_func_id: u32::MAX as u64,
                 closure: std::ptr::null_mut(),
                 defining_class: std::ptr::null_mut(),
             });
